POST Exercise Record and Exercise Bank with one API route
