+++ .gitignore
@@ -1,5 +1,6 @@
 # Notes
 dan-notes.md
+db/oldschema.sql
 
 # Logs
 logs
