--- controllers/api/exerciseRecordRoutes.js
@@ -1,6 +1,0 @@
-// Write new exercise record - make sure it has all the data?
-
-// Read an exercise record:
-// Check unit against user unit (1); convert if not
-// Read all values
-// Convert all values into javascript object.
