+++ controllers/api/index.js
@@ -1,12 +1,12 @@
 const router = require('express').Router();
 const userRoutes = require('./userRoutes');
 const workoutRoutes = require('./workoutRoutes');
-const bankRoutes = require('./bankRoutes')
+const exerciseRoutes = require('./exerciseRoutes')
 
 // Workout Routes, Exercise Routes, History Routes (?)
 
 router.use('/users', userRoutes);
 router.use('/workout', workoutRoutes);
-router.use('/bank', bankRoutes);
+router.use('/exercise', exerciseRoutes);
 
 module.exports = router;
