+++ seeds/exerciseRecordData.json
@@ -3,7 +3,7 @@
         "sets": 3,
         "unit": "kg",
         "workout_id": 1,
-        "exercise_id": 1,
+        "bank_id": 1,
         "reps1": 10,
         "reps2": 10,
         "reps3": 10,
@@ -15,7 +15,7 @@
         "sets": 4,
         "unit": "kg",
         "workout_id": 1,
-        "exercise_id": 2,
+        "bank_id": 2,
         "reps1": 8,
         "reps2": 10,
         "reps3": 12,
@@ -29,7 +29,7 @@
         "sets": 4,
         "unit": "kg",
         "workout_id": 1,
-        "exercise_id": 3,
+        "bank_id": 3,
         "reps1": 8,
         "reps2": 10,
         "reps3": 12,
@@ -43,7 +43,7 @@
         "sets": 3,
         "unit": "kg",
         "workout_id": 1,
-        "exercise_id": 1,
+        "bank_id": 1,
         "reps1": 10,
         "reps2": 10,
         "reps3": 10,
